Enable use of a round model
Previous design did not include a round model, there the matchups in the tournament only consisted of a List of a List of matchups.
Adding a RoundModel enables tracking of rounds by a specific object that by computing from the List of lists.
In addition, it becomes easier to activate or deactivate a round that has been completed during the tournament.